finished making lloyd collapsible, deleted lloydchat.jsand styled the collapsed bar temporarily
