+++ src/css/ChatWindow.css
@@ -1,16 +1,21 @@
 .botui-container {
   display: block;
-  margin: 2em auto;
+  /* margin: 2em auto; */
+  margin-left: auto;
+  margin-right: auto;
   width: 1000px;
   background-color: white;
   opacity: .9;
   border-radius: 10px;
+  background-color: var(--offwhite);
   box-shadow: 0px 0px 10px 1px var(--purple);
 }
 
 .botui-app-container {
-  margin-top: 80px;
-  width: 100vw;
+  /* margin-top: 80px; */
+  width: 1000px;
+  margin-left: auto;
+  margin-right: auto;
 }
 
 .botui-message-content {
