+++ src/css/LloydChat.css
@@ -1,4 +1,4 @@
-#lloyd-outline {
+/* #lloyd-outline {
   width: 100vw;
   position: relative;
 }
@@ -39,7 +39,7 @@
   right: 235px;
   width: 50px;
   height: 50px;
-}
+} */
 
 /* .botui-container {
   float: right;
