+++ src/css/MovieFeedItem.css
@@ -5,7 +5,7 @@
 /* ======================================================= */
 
 .movie-feed {
-  margin-top: 80px;
+  margin-top: 20px;
   display: flex;
   flex-direction: column;
   align-items: center;
